TrungNQ update code BE

+++ C0822G1-primaryschool-BE/src/main/java/com/c0822g1primaryschoolbe/entity/clazz/IClazzTeacher.java
@@ -1,9 +1,9 @@
 package com.c0822g1primaryschoolbe.entity.clazz;
 
 public interface IClazzTeacher {
-    int getId();
+    Long getId();
     String getName();
     String getSchoolYear();
-    int getTeacherId();
+    Long getTeacherId();
     String getTeacherName();
 }
